+++ src/dagon/core/application.d
@@ -1174,6 +1174,7 @@ class Application: EventListener, Updateable
         logInfo("GL_ARB_texture_compression_bptc: ", isExtensionSupported("GL_ARB_texture_compression_bptc"));
         
         // Initial OpenGL state settings
+        glPixelStorei(GL_PACK_ALIGNMENT, 1);
         glPixelStorei(GL_UNPACK_ALIGNMENT, 4);
         glClearColor(0.0f, 0.0f, 0.0f, 1.0f);
         glClearDepth(1.0);